Create a waffle component
