+++ src/pages/representatives.js
@@ -1,9 +1,8 @@
 import React from "react"
-import { graphql, Link } from "gatsby"
+import { graphql } from "gatsby"
 
 import Layout from "../components/layout"
 import SEO from "../components/seo"
-import "./waffle.css"
 
 export const query = graphql`
   query {
